+++ Tests/DangerTests/BitBucketServerTestResources/BitBucketServerUser.swift
@@ -1,6 +1,37 @@
 
 public let BitBucketServerUserJSON = """
+{
+    "name":"tum-id",
+    "emailAddress":"user.name@tum.de",
+    "id":1,
+    "displayName":"Firstname Lastname",
+    "active":true,
+    "slug":"tum-id",
+    "type":"NORMAL",
+    "links": {
+        "self": [
+            {
+                "href":"https://repobruegge.in.tum.de/users/tum-id"
+            }
+        ]
+    }
+}
+"""
 
-    {"name":"tum-id","emailAddress":"user.name@tum.de","id":1,"displayName":"Firstname Lastname","active":true,"slug":"tum-id","type":"NORMAL","links":{"self":[{"href":"https://repobruegge.in.tum.de/users/tum-id"}]}}
-
+public let BitBucketServerUserWithoutEmailJSON = """
+{
+    "name":"tum-id",
+    "id":1,
+    "displayName":"Firstname Lastname",
+    "active":true,
+    "slug":"tum-id",
+    "type":"NORMAL",
+    "links": {
+        "self": [
+            {
+                "href":"https://repobruegge.in.tum.de/users/tum-id"
+            }
+        ]
+    }
+}
 """