+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ## Master
 
+- Make BitBucketServerUser emailAddress optional [@f-meloni][] - [#257](https://github.com/danger/swift/pull/257)
+
 ## 1.6.4
 
 - Optimise Swiftlint perfomances [@f-meloni][] - [#256](https://github.com/danger/swift/pull/256)